feat: Add script to download files from S3 directory (#19)
The commit adds a new script, `download_s3_directory.py`, which allows users to download files from a specified S3 directory to a local directory. The script takes in the bucket name, S3 directory, and local directory as command line arguments and uses the `boto3` library to interact with the S3 API.

This script will be useful for users who need to download multiple files from an S3 directory in a convenient and automated way.